more work on linked notes
